requirements.txt
pytest == 8.0.2
requests
allure-pytest
selene>=2.0.0rc9